requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.62.3
httplib2==0.31.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pyparsing==3.2.5
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
